Create spectral_extraction_Nov2022.py (#303)
* Create spectral_extraction_Nov2022.py 

Script to extract spectra towards cores MM1 and MM2 in Sgr C. They should be in 36j in the mosaic and we will need 12m data -->  uid://A001/X15a0/X172. As a region I choose 0.75'' of radius to extract the spectra from about the 12m beam size. I have prepared a script based on that used in the past to extract other spectra. I am not sure about the path used since it starts from "rawdata". Adam could you please check if it is ok or if I should use an updated one with the reduced data (I see that pipelines have been run for this part of the 12 mosaic)?

* use the reduced data rather than the original delivered cubes

* correct typos in path

Co-authored-by: Adam Ginsburg <keflavich@gmail.com>