Updating tclean parameter for region ao 7M
Updating tclean parameter for Sgr_A_st_ao_03_7M, spw24: cyclefactor from 1.0 to 2.0 (issue #6)
But please note that I have run the reimaging on a local machine and have uploaded the images to Globus. This commit is just for bookkeeping, there is no need to rerun the pipeline on hipergator.